requirement.txt
Pillow>=10.0.0
imagehash>=4.3.1
pandas>=2.0.0
tqdm>=4.65.0
scikit-image>=0.21.0
numpy>=1.24.0